refactor onTutorialContinueConfig
