+++ src/actions/index.ts
@@ -1,4 +1,5 @@
 export { default as onStartup } from './onStartup'
 export { default as onTutorialConfig } from './onTutorialConfig'
+export { default as onTutorialContinueConfig } from './onTutorialContinueConfig'
 export { default as onErrorPage } from './onErrorPage'
 export { default as onTestPass } from './onTestPass'
